Adds a `pulumi new` command to scaffold a project
This adds a `pulumi new` command which makes it easy to quickly
automatically create the handful of needed files to get started building
an empty Pulumi project.

Usage:

```
$ pulumi new typescript
```

Or you can leave off the template name, and it will ask you to choose
one:

```
$ pulumi new
Please choose a template:
> javascript
  python
  typescript
```

By default it will use the name of the current working directory as the
project name and a description from the template.

The list of available templates are on the (currently private)
pulumi-templates org on GitHub. The repos in pulumi-templates are not
public yet. To access them I'll need to give you read access to the
repos and you'll need to set `PULUMI_GITHUB_TOKEN` to a GitHub personal
access token. Once we make the repos public, it won't be necessary to
set `PULUMI_GITHUB_TOKEN`.

Templates are cached under `~/.pulumi/templates`. When you use a
template that already exists, it will do a pull to ensure you have the
latest bits.

You can also have your own local templates under
`~/.pulumi/templates/<yourtemplate>`, which `pulumi new` will use if the
template cannot be found on GitHub.

If you're offline, `pulumi new` will allow the use of any templates that
exist under `~/.pulumi/templates`.