Adds native support for TypeScript to the Node.JS SDK
This allows `.ts` files to be loaded directly during `pulumi preview` and `pulumi update` operatrions - without the need to compile them to `.js` first.

The previous model of compiling separately is still an option, and provides maximal flexibility over the build toolchain.

But for the simple 80% case, this allows TypeScript to be used directly, with no need to do `tsc` or `yarn build` steps, and with no generated `bin` folder.