Error if importing cross-stack checkpoints
Per pulumi/pulumi#984, we will now issue an error if it appears you're
importing a checkpoint from a different stack.  This can be overridden
if you know what you're doing (with --force), but in general this is a
sign that you're doing something very wrong that will be hard to undo.