Require users to login
This changes the default behavior of the CLI to require that users
log into the Pulumi Cloud in order to perform deployments.  This takes
advantage of our new service work to act as a state and concurrency
management service, so that users don't need to do this in ad-hoc ways.