Add support for managing ICOs with Pulumi
As part of exploring Pulumi's next round of financing, we are
considering the possibility of issuing a new coin, CloudCoin, that
is redeemable for actual serverless compute in the cloud of your
choice.  We think this is the most expedient path to unicorn status.

To support this, we are simply adding a new resource provider
interface type, Coin, that may be implemented for issuing ICOs and
performing various useful maintenance on existing coin offerings.

A more comprehensive desing document for this feature can be found
at https://goo.gl/kxsioC.  Feedback is welcome!