Download Node v6.10.2 on install of pulumi/pulumi
This helps alleviate some of the pain of our system requiring that you
have a specific version of Node installed. With this change, we will
download the exact version of Node when the pulumi/pulumi npm package is
installed.

This change also places the native Node module inside of the package,
which allows us to version it in lockstep with the package and not as
part of the CLI.